Change Todo.categoryId type from Long to Category.categoryId
この変更に伴い、
todo表示画面に遷移する前に前カテゴリを取得し引数として渡す処理を追加
todo-itemに対応するカテゴリのslugを表示するように変更
todo追加、更新フォームではまだLongを入力する必要がある

+++ app/lib/persistence/Category.scala
@@ -80,7 +80,7 @@ case class CategoryRepository[P <: JdbcProfile]()(implicit val driver: P)
         // カテゴリを削除するクエリ
         val deleteCategory = categorySlick.filter(_.id === id).delete
         // 削除するカテゴリに紐づけられているtodoを更新するクエリ
-        val updateTodos    = todoSlick.filter(_.categoryId === id.toLong).map(_.categoryId).update(Category.deletedCategoryId)
+        val updateTodos    = todoSlick.filter(_.categoryId === id).map(_.categoryId).update(Category.deletedCategoryId)
         // 二つのクエリをトランザクション処理する
         db.run((deleteCategory andFinally updateTodos).transactionally)
       }(Predef.identity)
